GRE-404 Making sure that all unit tests extend the TestLog4JSetup trait
We need to set the system properties for the log filenames referred to by the log4j.xml configuration before any static Logger is loaded. We need to do this for more than just our own code, e.g. when a Logger is used by a library we rely on. Therefore please do make sure that you FIRST extend this trait and only then extend any other testing support and/or scalatra or jetty support classes from your unit tests.